Add files via upload
This is a perfect game to show off a little more about what exactly Python can do. Human input, comparisons, random selections and a whole host of loops are used in making a working version of the game. It is also easy enough to adapt and expand as you see fit, adding rules and results, and even making a rudimentary AI if you wish.