Add files via upload
An interesting German mathematician formula based on the following rules:
Take any natural number. There is a rule, or function, which we apply to that number, to get the next number. We then apply that rule over and over, and see where it takes us. The rule is this: If the number is even, then divide it by 2, and if the number is odd, then multiply by 3 and add 1.